domd/domu: Remove virtio.scc to fix Linux build issue
During Linux build Yocto complains that input file "cfg/virtio.scc"
does not exist
This is because there is no yocto-kernel-cache recipe which provides
such configuration.
https://lists.yoctoproject.org/g/meta-virtualization/topic/linux_5_4_input_file/74940175?p=

Remove virtio.scc for now to allow Linux build to complete,
but it is not going to be an issue since all related virtio
configs are already enabled in Linux defconfig.

TODO: Examine if virtio.scc is necessary in current environment
and consider including yocto-kernel-cache recipe.

Signed-off-by: Valerii Chubar <valerii_chubar@epam.com>